micro-balloon: Support Block API Version 2
