+++ blocks/src/micro/micro-balloon/block.json
@@ -1,0 +1,46 @@
+{
+  "apiVersion": 2,
+  "name": "cocoon-blocks/micro-balloon-2",
+  "category": "cocoon-micro",
+  "keywords": [ "micro", "copy", "balloon" ],
+
+  "attributes": {
+    "content": {
+      "type": "string",
+      "default": "マイクロコピーバルーン"
+    },
+    "type": {
+      "type": "string",
+      "default": "micro-top"
+    },
+    "isCircle": {
+      "type": "boolean",
+      "default": false
+    },
+    "icon": {
+      "type": "string"
+    },
+    "align": {
+      "type": "string"
+    },
+    "backgroundColor": {
+      "type": "string"
+    },
+    "textColor": {
+      "type": "string"
+    },
+    "borderColor": {
+      "type": "string"
+    },
+    "customBorderColor": {
+      "type": "string"
+    },
+    "fontSize": {
+      "type": "string"
+    }
+  },
+  "supports": {
+    "align": [ "left", "center", "right" ],
+    "customClassName": true
+  }
+}
